requirements.txt
fastapi==0.109.0
pillow==10.1.0
python-multipart==0.0.6
uvicorn==0.25.0
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.3
typing-extensions==4.9.0
starlette==0.35.1
jinja2==3.1.2
html2image==2.1.1